move/tidy up a bit more
