+++ test/__nodetests/basic.js
@@ -1,5 +1,7 @@
-var Ractive = require( '../../ractive' ),
-	assert = require( 'assert' );
+/*global require, describe, it */
+
+var Ractive = require( '../../ractive' );
+var assert = require( 'assert' );
 
 describe( 'Ractive', function () {
 	it( 'should be a function', function () {
