+++ test/__tests/arrayMethods.js
@@ -1,160 +1,3 @@
-var baseItems = [ 'alice', 'bob', 'charles' ];
-
-[ true, false ].forEach( modifyArrays => {
-	var List = Ractive.extend({
-		template: '<ul>{{#items}}<li>{{.}}</li>{{/items}}</ul>',
-		modifyArrays: modifyArrays
-	});
-
-	test( 'ractive.push() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.push( 'items', 'dave' );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>alice</li><li>bob</li><li>charles</li><li>dave</li></ul>' );
-	});
-
-	asyncTest( 'ractive.pop() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-		expect(2);
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.pop( 'items' ).then( function(v) {
-			t.strictEqual( v, 'charles' );
-			QUnit.start();
-		} );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>alice</li><li>bob</li></ul>' );
-	});
-
-	asyncTest( 'ractive.shift() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-		expect(2);
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.shift( 'items' ).then( function(v) {
-			t.strictEqual( v, 'alice' );
-			QUnit.start();
-		} );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>bob</li><li>charles</li></ul>' );
-	});
-
-	test( 'ractive.unshift() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.unshift( 'items', 'dave');
-		t.htmlEqual( fixture.innerHTML, '<ul><li>dave</li><li>alice</li><li>bob</li><li>charles</li></ul>' );
-	});
-
-	asyncTest( 'ractive.splice() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-		expect(5);
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.splice( 'items', 1, 1, 'dave', 'eric' ).then( function(v) {
-			t.deepEqual( v, [ 'bob' ] );
-			QUnit.start();
-		} );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>alice</li><li>dave</li><li>eric</li><li>charles</li></ul>' );
-
-		// removing before the beginning removes from the beginning
-		ractive.splice( 'items', -10, 1, 'john' );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>john</li><li>dave</li><li>eric</li><li>charles</li></ul>' );
-
-		// removing beyond the end is a noop
-		ractive.splice( 'items', 10, 1, 'larry' );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>john</li><li>dave</li><li>eric</li><li>charles</li><li>larry</li></ul>' );
-
-		// negative indexing within bounds starts from the end
-		ractive.splice( 'items', -1, 1 );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>john</li><li>dave</li><li>eric</li><li>charles</li></ul>' );
-	});
-
-	test( 'ractive.reverse() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.reverse( 'items' );
-		t.htmlEqual( fixture.innerHTML, '<ul><li>charles</li><li>bob</li><li>alice</li></ul>' );
-	});
-
-	test( 'ractive.sort() (modifyArrays: ' + modifyArrays + ')', function ( t ) {
-		var items, ractive;
-
-		items = baseItems.slice();
-
-		ractive = new List({
-			el: fixture,
-			data: { items: items }
-		});
-
-		ractive.sort( 'items', function ( a, b ) {
-			return a.length - b.length;
-		});
-		t.htmlEqual( fixture.innerHTML, '<ul><li>bob</li><li>alice</li><li>charles</li></ul>' );
-	});
-});
-
-asyncTest( 'Array method proxies return a promise that resolves on transition complete', function ( t ) {
-	var items, ractive;
-
-	items = baseItems.slice();
-
-	ractive = new Ractive({
-		el: fixture,
-		template: '<ul>{{#items}}<li intro="test">{{.}}</li>{{/items}}</ul>',
-		data: { items: items },
-		transitions: {
-			test: function ( t ) {
-				setTimeout( t.complete, 50 );
-			}
-		}
-	});
-
-	expect( 1 );
-
-	ractive.push( 'items', 'dave' ).then( function () {
-		t.htmlEqual( fixture.innerHTML, '<ul><li>alice</li><li>bob</li><li>charles</li><li>dave</li></ul>' );
-		QUnit.start();
-	});
-});
-
 test( 'Pattern observers on arrays fire correctly after mutations (mirror of test in observe.js)', t => {
 	const ractive = new Ractive({
 		data: {
@@ -276,7 +119,7 @@ test( 'array modification with non-shuffle-able deps should update correctly', t
 	t.htmlEqual( fixture.innerHTML, '0120' );
 });
 
-test( 'shifting an empty array', t => {
+test( 'shifting an empty array', () => {
 	expect( 0 );
 
 	let ractive = new Ractive({
